Use "netsh firewall" to allow pythonw.exe to execute
By default, the Windows Firewall prevents newly-installed applications
from a network share from executing without a user confirmation. This
prevents the cuckoo agent from running. Disable the behavior for
pythonw.exe for now, but we may want to change this in the future to
simply turn off the Windows Firewall and Windows Defender services
altogether.